Fixing logging configuration
